+++ backend/jobs/import/intercarif/utils/getFormationsFromCSV.js
