+++ backend/test/integration/jobs/import/accounts/intercarifAccountImporter-test.js
@@ -3,7 +3,7 @@ const assert = require('assert');
 const { withMongoDB } = require('../../../../helpers/test-db');
 const { newOrganismeAccount, newComment } = require('../../../../helpers/data/dataset');
 const logger = require('../../../../helpers/test-logger');
-const intercarifOrganismesImporter = require('../../../../../jobs/import/account/importers/intercarifAccountImporter');
+const intercarifOrganismesImporter = require('../../../../../jobs/import/accounts/importers/intercarifAccountImporter');
 
 describe(__filename, withMongoDB(({ getTestDatabase, insertIntoDatabase }) => {
