+++ backend/test/integration/jobs/import/organismes/generateOrganismesFormateurs-test.js
@@ -2,7 +2,7 @@ const assert = require('assert');
 const _ = require('lodash');
 const { withMongoDB } = require('../../../../helpers/test-db');
 const { newComment, newOrganismeResponsable } = require('../../../../helpers/data/dataset');
-const generateOrganismesFormateurs = require('../../../../../jobs/import/intercarif/steps/generateOrganismesFormateurs');
+const generateOrganismesFormateurs = require('../../../../../jobs/import/organismes/generateOrganismesFormateurs');
 
 describe(__filename, withMongoDB(({ getTestDatabase, insertIntoDatabase }) => {
