+++ backend/test/integration/jobs/import/organismes/generateOrganismesResponsables-test.js
@@ -1,6 +1,6 @@
 const assert = require('assert');
 const { withMongoDB } = require('../../../../helpers/test-db');
-const generateOrganismesResponsables = require('../../../../../jobs/import/intercarif/steps/generateOrganismesResponsables');
+const generateOrganismesResponsables = require('../../../../../jobs/import/organismes/generateOrganismesResponsables');
 
 describe(__filename, withMongoDB(({ getTestDatabase, importIntercarif }) => {
