+++ backend/test/unit/jobs/import/intercarif/getFormationsFromCSV-test.js
@@ -1,6 +1,6 @@
 const assert = require('assert');
 const path = require('path');
-const getFormations = require('../../../../../jobs/import/intercarif/steps/utils/getFormationsFromCSV');
+const getFormations = require('../../../../../jobs/import/intercarif/utils/getFormationsFromCSV');
 
 describe(__filename, () => {
